requirements.txt
dash
dash-html-components
biopython==1.78
py2neo==2020.0.0
